+++ dms/v_analysis_delta_t_report.sql
@@ -4,7 +4,7 @@
 
 CREATE VIEW public.v_analysis_delta_t_report AS
  SELECT j.job,
-    (EXTRACT(epoch FROM (j.finish - j.start)) / 60.0) AS delta_t,
+    (EXTRACT(epoch FROM (j.finish - j.start)) / (60)::numeric) AS delta_t,
     j.priority,
     js.job_state AS state,
     tool.analysis_tool AS tool_name,
